requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
python-dotenv
PyMySQL
Werkzeug
deepface
Pillow
tensorflow==2.12.0
cryptography
Flask-JWT-Extended
